feat(conf): improve OpenAPI\Swagger documentation
Resolve MON-3235
