+++ src/CentreonModule/Application/Webservice/CentreonModuleWebservice.php
@@ -7,7 +7,7 @@ use Centreon\Application\DataRepresenter\Response;
 use CentreonModule\Application\DataRepresenter\ModuleEntity;
 
 /**
- * @OA\Tag(name="centreon_module", description="")
+ * @OA\Tag(name="centreon_module", description="Resource for authorized access")
  */
 class CentreonModuleWebservice extends CentreonWebServiceAbstract
 {
@@ -138,102 +138,6 @@ class CentreonModuleWebservice extends CentreonWebServiceAbstract
     }
 
     /**
-     * @OA\Post(
-     *   path="/external.php?object=centreon_module&action=getBamModuleInfo",
-     *   summary="Get list of modules and widgets",
-     *   tags={"centreon_module"},
-     *   @OA\Parameter(
-     *       in="query",
-     *       name="object",
-     *       description="the name of the API object class",
-     *       required=true,
-     *       @OA\Schema(
-     *          type="string",
-     *          enum={"centreon_modules_webservice"}
-     *       )
-     *   ),
-     *   @OA\Parameter(
-     *       in="query",
-     *       name="action",
-     *       description="the name of the action in the API class",
-     *       required=true,
-     *       @OA\Schema(
-     *          type="string",
-     *          enum={"getBamModuleInfo"}
-     *       )
-     *   ),
-     *   @OA\Response(
-     *      response="200",
-     *      description="JSON with BAM module info",
-     *      @OA\MediaType(
-     *          mediaType="application/json",
-     *          @OA\Schema(
-     *              @OA\Property(property="enabled", type="boolean"),
-     *              @OA\Property(property="status", type="boolean"),
-     *          )
-     *      )
-     *   )
-     * )
-     * @OA\Post(
-     *   path="/external.php?object=centreon_modules_webservice&action=getBamModuleInfo",
-     *   summary="Get list of modules and widgets",
-     *   tags={"centreon_modules_webservice"},
-     *   @OA\Parameter(
-     *       in="query",
-     *       name="object",
-     *       description="the name of the API object class",
-     *       required=true,
-     *       @OA\Schema(
-     *          type="string",
-     *          enum={"centreon_modules_webservice"}
-     *       )
-     *   ),
-     *   @OA\Parameter(
-     *       in="query",
-     *       name="action",
-     *       description="the name of the action in the API class",
-     *       required=true,
-     *       @OA\Schema(
-     *          type="string",
-     *          enum={"getBamModuleInfo"}
-     *       )
-     *   ),
-     *   @OA\Response(
-     *      response="200",
-     *      description="JSON with BAM module info",
-     *      @OA\MediaType(
-     *          mediaType="application/json",
-     *          @OA\Schema(
-     *              @OA\Property(property="enabled", type="boolean"),
-     *              @OA\Property(property="status", type="boolean"),
-     *          )
-     *      )
-     *   )
-     * )
-     *
-     * Get info for BAM module
-     * 
-     * @return array
-     */
-    public function postGetBamModuleInfo(): array
-    {
-        $factory = new \CentreonLegacy\Core\Utils\Factory($this->getDi());
-        $utils = $factory->newUtils();
-        $moduleFactory = new \CentreonLegacy\Core\Module\Factory($this->getDi(), $utils);
-        $moduleInfoObj = $moduleFactory->newInformation();
-        $modules = $moduleInfoObj->getList();
-
-        if (
-            array_key_exists('centreon-bam-server', $modules) &&
-            $modules['centreon-bam-server']['is_installed']
-        ) {
-            return ['enabled' => true];
-        }
-
-        return ['enabled' => false];
-    }
-
-    /**
      * Authorize to access to the action
      *
      * @param string $action The action name
@@ -244,7 +148,11 @@ class CentreonModuleWebservice extends CentreonWebServiceAbstract
      */
     public function authorize($action, $user, $isInternal = false)
     {
-        return true;
+        if (parent::authorize($action, $user, $isInternal)) {
+            return true;
+        }
+
+        return $user && $user->hasAccessRestApiConfiguration();
     }
 
     /**
